Divide User API into Use API and UserWords API
