+++ BusinessLayer/Services/Contracts/IUserWordsService.cs
@@ -1,0 +1,18 @@
+﻿
+
+using System.Collections.Generic;
+using UrbanDictionary.BusinessLayer.DTO;
+
+namespace UrbanDictionary.BusinessLayer.Services.Contracts
+{
+    public interface IUserWordsService
+    {
+        public IEnumerable<WordDTO> GetSavedWords();
+        public bool TryAddToSavedWords(long id);
+        public bool TryDeleteSavedWord(long id);
+
+        public IEnumerable<WordDTO> GetCreatedWords();
+
+        public void Save();
+    }
+}
