+++ UrbanDictionary/Controllers/UsersController.cs
@@ -20,61 +20,11 @@ namespace UrbanDictionary.Controllers
             _serviceWrapper = serviceWrapper;
         }
 
-        [HttpGet("savedWords")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult<IEnumerable<WordDTO>> GetSavedWords()
-        {
-            var savedWords = _serviceWrapper.User.GetSavedWords();
-            if (savedWords != null)
-            {
-                return Ok(savedWords);
-            }
-            return NotFound();
-        }
-
-        [HttpPost("saveWord")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public ActionResult SaveWord(long id)
-        {
-            if (_serviceWrapper.User.TryAddToSavedWords(id))
-            {
-                return Ok();
-            }
-            return BadRequest();
-        }
-
-        [HttpDelete("deleteSavedWord/{id}")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult DeleteSavedWords(long id)
-        {
-            if (_serviceWrapper.User.TryDeleteSavedWord(id))
-            {
-                return Ok();
-            }
-            return NotFound();
-        }
-
-        [HttpGet("createdWords")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult<IEnumerable<WordDTO>> GetCreatedWords()
-        {
-            var createdWords = _serviceWrapper.User.GetCreatedWords();
-            if (createdWords != null)
-            {
-                return Ok(createdWords);
-            }
-            return NotFound();
-        }
-
         [HttpGet]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public ActionResult<IEnumerable<WordDTO>> Get()
         {
-            return Ok(_serviceWrapper.User.GetUsers());
+            return Ok(_serviceWrapper.User.GetAll());
         }
 
 
@@ -83,7 +33,7 @@ namespace UrbanDictionary.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public ActionResult Delete(string id)
         {
-            if (_serviceWrapper.User.TryDeleteUser(id))
+            if (_serviceWrapper.User.TryDelete(id))
             {
                 return Ok();
             }
